Fix slowest rules report from --time flag (#11066)
The "slowest rules" report which is printed when `--time` is passed to
`semgrep scan` was not printing the slowest rules. Instead, it was
printing the last five rules, when the list is sorted alphabetically.

This change ensures that the rules are sorted by run-time instead of
rule name.

Fixes https://github.com/semgrep/semgrep/issues/11065

(cc @samueljsb)

Co-authored-by: Samuel Searles-Bryant <sam@samueljsb.co.uk>