Revert "refactor: always call `Eio_main.run` (semgrep/semgrep-proprietary#3850)" (semgrep/semgrep-proprietary#4189)
This reverts commit 7a63af29ced77115df741eea1c30047cf46131a7.

We believe this patch contributed to INCID-208, where scans were failing
owing to unhandled `Timeout` exceptions _inside_ an effect handler.
Since these scans were still using Parmap it was surprising to see an
effect handler doing much of anything, least of all raising an exn like
this.

What we believe happened was this: during the processing of an effect
like handling a simple yield or getting a fiber-local value, inside a
computation with some timeout set, the SIGALRM fires, and _whilst inside
the effect handler_ we execute the signal handler; the latter raises
`Timeout` and control returns back to where it was before the signal
fired: inside the effect handler.  However, this is not in the continuation 
(in `Time_limit.ml`) where we intended to catch that exn.  As a result, 
the `Timeout` is ultimately not handled at any level, and the process
 exits with error code 2.

Coupled with the above is a refactor of #4090, which depended on the
patch we are backing out, to explicitly start the Eio event loop for the
LSP server within `Lsp_subcommand.main` and the LSP's test suite,
and to remove extraneous `base` (technically the eio environment)
arguments from the unit test suite.

Previously, setting up Eio was handled automatically by `Eio_main`
unconditionally starting up.  Splitting the revert and the refactor
would have temporarily left the LSP broken.

synced from Pro e47eb3810567b0cdcad421acd38c055695d6aa73