fix: Add an InvalidNotInOr invalid_rule_kind (semgrep/semgrep-proprietary#4301)
Currently, the OCaml rule parser doesn't check for uses of Not inside an
Or and we rely on the validation of rules to catch it. This commit adds
an additional check to catch this as an invalid rule.

Test plan:

- Un-quarantines a test that was previously quarantined when the OCaml
parser based RPC validation was added to short-circuit the slow Python
JSON schema validation in the happy case.

synced from Pro c238671637963e04dfbb1f3e5aa9dc7a3fd0f43a