fix(parallelism): threadsafe =~ memoization (semgrep/semgrep-proprietary#4770)
We had previously observed a race in Common.candidate_match_func owing
to racing on a shared memoization cache.

The original plan was to migrate this memoization to the standard
[SharedMemo] abstraction, which synchronises a single cache against all
threads. However, we couldn't do so to fix this race right away because
we needed to disentangle the parallelism library from common so the
latter could take a dependency on the former. This was done as of
54d1811e9ad2198d3940951e919bcbb9c3f82b68 - however, the followup work to
fix the race didn't actually happen.

This patch originally did so; however, despite all evidence to the
contrary[1], `regexp` appears to be racy and TSAN started reporting data
races: E.g.:

```
WARNING: ThreadSanitizer: data race (pid=74315)
  Read of size 1 at 0x00011aacba60 by thread T1 (mutexes: write M0):
    #0 re_match strstubs.c:217 (Main.exe:arm64+0x103e3b60c)
    #1 re_string_match strstubs.c:407 (Main.exe:arm64+0x103e3b238)
    #2 caml_c_call <null> (Main.exe:arm64+0x1044c2ca8)
    #3 camlStr.string_match_1297 <null> (Main.exe:arm64+0x103c32a30)
...

  Previous atomic write of size 8 at 0x00011aacba60 by thread T10 (mutexes: write M1):
    #0 caml_alloc_string alloc.c:188 (Main.exe:arm64+0x104471ae0)
    #1 caml_create_bytes str.c:78 (Main.exe:arm64+0x1044b8cbc)
    #2 caml_c_call <null> (Main.exe:arm64+0x1044c2ca8)
    #3 camlStdlib__Bytes.sub_305 <null> (Main.exe:arm64+0x103d521ac)
    #4 camlStr.flush_1163 <null> (Main.exe:arm64+0x103c3020c)
...
```

I am quite sure that these races are benign (writes are atomic even
though the read is not, and the write was during data structure
initialisation), but to be conservative, this patch instead uses a
per-domain hash table for memoisation rather than the global one. This
does mean very common regexps will be duplicated across domains, but
perf and memory jobs don't seem to indicate that this is an issue.

--

[1] for the curious: https://github.com/ocaml/ocaml/issues/8091 on the
OCaml issue tracker talk about "re_group" and "re_register" as
problematic, but those appear to have been moved onto the stack as of
commit 45343f6a39c . The only other thing that stood out was the use of
the volatile keyword, but that appears to be an artefact of making field
accesses volatile in general (in 0e09dc477c1). Broadly, the only
occurrences of Field as an lval appear on re_alloc_groups:138-142, on a
locally-allocated heap value.

Test plan: CI, no longer observe this particular race whilst running
under TSAN.

synced from Pro 682a5aa7682c65d431a42303a8b324f8d78c7579