feat: safe executor pool (semgrep/semgrep-proprietary#4878)
Eio's executor pool does not recover if there is an uncaught exception in the domain. This sometimes happens when an async exception is delivered in eio scheduling code.

This PR fixes this by forking executor pool specifically to resolve promises with the uncaught exception if it occurs, and then re-adds a domain to the pool.

Test plan:
I added a unit test that triggers the type of exception we're worried about, and ensures its caught and domains restart.

CI passes

synced from Pro 4e7ea52ba4f5cc499b6c825bc2ee92a50021fd3b